Fixed member and added memv, also fixed t() to use is_equal to compare results.
member was using === as the default equality test, therefore acting like memv.
member was changed to use is_equal as the default.
Tests were added that use Pairs and arrays as keys (that wouldn't compare ===).
t() was using is_equal only when the expected result was a pair.
It now uses is_equal for all tests.